Memory optimization

+++ implementation/Abstraction.cpp
@@ -8,21 +8,22 @@
 #include <algorithm>
 #include <iostream>
 #include <ranges>
+#include <cstdint>
 
 namespace topspin {
 
-using State = std::vector<int>;
+using State = std::vector<uint8_t>;
 struct VecHash {
     size_t operator()(const State& v) const noexcept {
         size_t h = 0;
-        for (int x : v) h ^= std::hash<int>{}(x) + 0x9e3779b9 + (h << 6) + (h >> 2);
+        for (uint8_t x : v) h ^= std::hash<uint8_t>{}(x) + 0x9e3779b9 + (h << 6) + (h >> 2);
         return h;
     }
 };
 static std::unordered_map<State, int, VecHash> solutionLengthCache;
 
-std::vector<int> abstract_state(const std::vector<int>& input, const std::function<bool(const int&)>& predicate) {
-    std::vector<int> abstraction;
+std::vector<uint8_t> abstract_state(const std::vector<uint8_t>& input, const std::function<bool(const uint8_t&)>& predicate) {
+    std::vector<uint8_t> abstraction;
     abstraction.reserve(input.size());
     for (const auto& item : input) {
         abstraction.push_back(predicate(item) ? item : 0);
@@ -30,33 +31,29 @@ std::vector<int> abstract_state(const std::vector<int>& input, const std::functi
     return abstraction;
 }
 
-std::vector<int> abstract_stateC(const std::vector<int>& input, const std::function<int(const int&)>& mapping) {
-    std::vector<int> abstraction(input.size());
+std::vector<uint8_t> abstract_stateC(const std::vector<uint8_t>& input, const std::function<int(const uint8_t&)>& mapping) {
+    std::vector<uint8_t> abstraction(input.size());
     std::ranges::transform(input, abstraction.begin(), mapping);
     return abstraction;
 }
 
-//Normalize the vector by shifting elements to relative positions of smallest non-zero element
-std::vector<int> normalize(std::vector<int> state) {
-    auto filtered = state | std::views::filter([](int x) { return x != 0; });
-    int min_non_zero = *std::ranges::min_element(filtered);
+std::vector<uint8_t> normalize(std::vector<uint8_t> state) {
+    auto filtered = state | std::views::filter([](uint8_t x) { return x != 0; });
+    uint8_t min_non_zero = *std::ranges::min_element(filtered);
     
     auto it = std::ranges::find(state, min_non_zero);
     int current_index = std::distance(state.begin(), it);
     int target_index = min_non_zero - 1;
 
-    // Compute rotation distance
     int shift = (target_index - current_index + state.size()) % state.size();
 
-    // Rotate to move min_non_zero to target_index
     std::rotate(state.begin(), state.begin() + (state.size() - shift), state.end());
     return state;
 }
 
-// Check if the abstraction is a goal state by normalizing it and checking if it matches the expected goal state
-bool is_goal(const std::vector<int>& abstraction) {
+bool is_goal(const std::vector<uint8_t>& abstraction) {
     const int n = static_cast<int>(abstraction.size());
-    std::vector<int> normalized = normalize(abstraction);
+    std::vector<uint8_t> normalized = normalize(abstraction);
     for (int i = 0; i < n; i++) {
         if (normalized[i] == 0) {
             continue;
@@ -67,17 +64,16 @@ bool is_goal(const std::vector<int>& abstraction) {
     return true;
 }
 
-bool is_goalC(const std::vector<int>& abstraction, const std::function<int(int)>& mapping) {
+bool is_goalC(const std::vector<uint8_t>& abstraction, const std::function<int(uint8_t)>& mapping) {
     const int n = static_cast<int>(abstraction.size());
-    // Build the expected goal pattern using the mapping
     for (int i = 0; i < n; i++) {
         if (abstraction[i] != mapping(i + 1)) return false;
     }
     return true;
 }
 
-std::vector<int> subvec_wraparound(const std::vector<int>& vec, int pos, int len) {
-    std::vector<int> result;
+std::vector<uint8_t> subvec_wraparound(const std::vector<uint8_t>& vec, int pos, int len) {
+    std::vector<uint8_t> result;
     int n = static_cast<int>(vec.size());
     result.reserve(len);
     for (int i = 0; i < len; i++) {
@@ -86,16 +82,16 @@ std::vector<int> subvec_wraparound(const std::vector<int>& vec, int pos, int len
     return result;
 }
 
-bool non_zero(const std::vector<int>& state, int pos, int n, int k) {
+bool non_zero(const std::vector<uint8_t>& state, int pos, int n, int k) {
     for (int i = 0; i < k; i++) {
         if (state[(pos + i) % n] != 0) return true;
     }
     return false;
 }
 
-std::vector<int> reverseWindow(const std::vector<int>& state, int pos, int k) {
+std::vector<uint8_t> reverseWindow(const std::vector<uint8_t>& state, int pos, int k) {
     const int n = static_cast<int>(state.size());
-    std::vector<int> newState = state;
+    std::vector<uint8_t> newState = state;
 
     for (int i = 0; i < k / 2; i++) {
         int left  = (pos + i) % n;
@@ -106,7 +102,7 @@ std::vector<int> reverseWindow(const std::vector<int>& state, int pos, int k) {
 }
 
 template<typename GoalFunc>
-int getSolutionLengthGeneric(const std::vector<int>& abstraction, int k, GoalFunc is_goal_func) {
+int getSolutionLengthGeneric(const std::vector<uint8_t>& abstraction, int k, GoalFunc is_goal_func) {
     State key = normalize(abstraction);
 
     auto it = solutionLengthCache.find(key);
@@ -147,12 +143,12 @@ int getSolutionLengthGeneric(const std::vector<int>& abstraction, int k, GoalFun
     return -1;
 }
 
-int getSolutionLength(const std::vector<int>& abstraction, int k) {
+int getSolutionLength(const std::vector<uint8_t>& abstraction, int k) {
     return getSolutionLengthGeneric(abstraction, k, is_goal);
 }
 
-int getSolutionLengthC(const std::vector<int>& abstraction, int k, const std::function<int(int)>& mapping) {
-    return getSolutionLengthGeneric(abstraction, k, [&](const std::vector<int>& state) {
+int getSolutionLengthC(const std::vector<uint8_t>& abstraction, int k, const std::function<int(uint8_t)>& mapping) {
+    return getSolutionLengthGeneric(abstraction, k, [&](const std::vector<uint8_t>& state) {
         return is_goalC(state, mapping);
     });
 }
